Support Fetching IMSI MCC and MNC for RuimRecords.
The Following changes are done with this.

- Read IMSI using SIM I/O for CSIM/RUIM.

Use Sim I/O command to read the EF_IMSIM to retrieve the IMSI
from the card, rather than using CommandsInterface.getIMSIForApp.
When both SIM and RUIM app exists, we cannot specify the
application for which we should retrieve the IMSI using the getIMSI
API, since SIM and RUIM have no AID.

- Change to return proper MNC in case of CSIM.

As per spec C.S0065 section 5.2.2 MNC length in case of CSIM
IMSI is 2. So modify the code to return the MNC of proper length.

- Introduce getOperatorNumeric api in RuimRecords

Since the DcTracker uses Generalized API
getOperatorNumeric for both SimRecords and
RuimRecords, Introduces same API in RuimRecords
which indirectly invoke getRuimOperatorNumeric

Change-Id: I905d4a271dfba168f2bc3830b8a50733cbe17f8c
CRs-fixed: 514270,656508
Signed-off-by: BijiOnta <sengolcaduk32@gmail.com>